C# + Some Node.js Multi-Module Work

+++ scripts/generate-raw-coverage.sh
@@ -2,25 +2,24 @@
 
 set -euxo pipefail
 
+if [[ ! -d integration-tests ]]; then
+    echo "Not in root of FishyJoes"
+    exit 1
+fi
+
 export GRADLE_OPTS=-Dorg.gradle.daemon=false
 
 export FISHYJOES_COVERAGE_PATH=$PWD/coverage-data
 rm -rf $FISHYJOES_COVERAGE_PATH
 mkdir -p $FISHYJOES_COVERAGE_PATH
 
-javaLibDir=kotlin-runtime/src/generated/resources/mac
-cSharpLibDir=c-sharp-runtime/runtimes/osx/native
-
-rm -rf $javaLibDir $cSharpLibDir
-mkdir -p $javaLibDir $cSharpLibDir
-
 # swift 5.7 no longer recognizes "--enable-code-coverage" outside of the "test" command
 COVERAGE_FLAGS=(-Xswiftc -profile-coverage-mapping -Xswiftc -profile-generate)
 
-swift build --configuration debug $COVERAGE_FLAGS --product FishyJoesJavaRuntime
-swift build --configuration debug $COVERAGE_FLAGS --product FishyJoesCSharpRuntime
-cp .build/debug/libFishyJoesJavaRuntime.dylib $javaLibDir
-cp .build/debug/libFishyJoesCSharpRuntime.dylib $cSharpLibDir
+CONFIGURATION=debug SKIP_LIPO=1 ./scripts/compile-node-runtime.sh $COVERAGE_FLAGS
+CONFIGURATION=debug SKIP_LIPO=1 ./scripts/compile-kotlin-native-runtime.sh $COVERAGE_FLAGS
+CONFIGURATION=debug SKIP_LIPO=1 ./scripts/compile-c-sharp-runtime.sh $COVERAGE_FLAGS
+
 (cd kotlin-runtime && ./gradlew publishToMavenLocal)
 
 # Gather coverage from kotlin tests
@@ -42,7 +41,7 @@ cp .build/debug/libFishyJoesCSharpRuntime.dylib $cSharpLibDir
 (
     cd integration-tests/TestAPI-bindings
     export LLVM_PROFILE_FILE=$FISHYJOES_COVERAGE_PATH/integration-tests-generate-build.profraw
-    swift run $COVERAGE_FLAGS -- fishy-joes generate build --kotlin-fast --nodejs --debug
+    swift run $COVERAGE_FLAGS -- fishy-joes generate build --kotlin-fast --nodejs --c-sharp --debug
 )
 
 (
@@ -54,13 +53,13 @@ cp .build/debug/libFishyJoesCSharpRuntime.dylib $cSharpLibDir
 (
     cd integration-tests/TestAPI-bindings
     export LLVM_PROFILE_FILE=$FISHYJOES_COVERAGE_PATH/integration-tests-kotlin.profraw
-    swift run $COVERAGE_FLAGS -- fishy-joes build test --kotlin-fast --debug
+    swift run $COVERAGE_FLAGS -- fishy-joes test --kotlin-fast --debug
 )
 
 (
     cd integration-tests/TestAPI-bindings
     export LLVM_PROFILE_FILE=$FISHYJOES_COVERAGE_PATH/integration-tests-c-sharp.profraw
-    swift run $COVERAGE_FLAGS -- fishy-joes build test --c-sharp --debug
+    swift run $COVERAGE_FLAGS -- fishy-joes test --c-sharp --debug
 )
 
 # Check that generation didn't change anything
